Added function to simulate environment data (issf).
Edited simulate_FM to support optional file ID to write some data
to file - this could be used by the web app to display data as it
is produced.

Added run_simulation function which is a different way to run the
simulation as used by the web app. It simulates periodic environment
data.